fix(schema): improve reason for UUID cast error, currently a TypeError
Fix #15202
